pyproject.toml
[project]
name = "metricas-studdeo-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.121.2",
    "pre-commit>=4.5.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.121.2",
    "pytest>=9.0.1",
]
